added code for limit and skip

+++ routers/post.py
@@ -14,13 +14,14 @@ router = APIRouter(
 
 
 @router.get("/", response_model=List[schemas.Post])
-async def get_posts(db: Session = Depends(get_db), current_user: str = Depends(oauth2.get_current_user)):
-    posts = db.query(models.Post).all()
+async def get_posts(db: Session = Depends(get_db), current_user=Depends(oauth2.get_current_user), limit: int = 10, skip: int = 0):
+    # posts = db.query(models.Post).filter(current_user.id == models.Post.owner_id).all()
+    posts = db.query(models.Post).limit(limit).offset(skip).all()
     return posts
 
 
 @router.get("/{id}", response_model=schemas.Post)
-async def get_post(id: int, db: Session = Depends(get_db), current_user: str = Depends(oauth2.get_current_user)):
+async def get_post(id: int, db: Session = Depends(get_db), current_user=Depends(oauth2.get_current_user)):
     post = db.query(models.Post).filter(models.Post.id == id).first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
@@ -29,8 +30,9 @@ async def get_post(id: int, db: Session = Depends(get_db), current_user: str = D
 
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.Post)
-async def create_post(post: schemas.PostCreate, db: Session = Depends(get_db), current_user: str = Depends(oauth2.get_current_user)):
-    new_post = models.Post(**post.model_dump())
+async def create_post(post: schemas.PostCreate, db: Session = Depends(get_db), current_user=Depends(oauth2.get_current_user)):
+    new_post = models.Post(owner_id=current_user.id, **
+                           post.model_dump())
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -38,22 +40,28 @@ async def create_post(post: schemas.PostCreate, db: Session = Depends(get_db), c
 
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(id: int,  db: Session = Depends(get_db), current_user: str = Depends(oauth2.get_current_user)):
+async def delete_post(id: int,  db: Session = Depends(get_db), current_user=Depends(oauth2.get_current_user)):
     post = db.get(models.Post, id)
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Post with id: {id} is not found")
+    if (current_user.id != post.owner_id):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f"Not authorized to delete this post")
     db.delete(post)
     db.commit()
     return post
 
 
 @router.put("/{id}", response_model=schemas.Post, status_code=status.HTTP_200_OK)
-async def update_post(id: int, post_data: schemas.PostCreate, db: Session = Depends(get_db), current_user: str = Depends(oauth2.get_current_user)):
+async def update_post(id: int, post_data: schemas.PostCreate, db: Session = Depends(get_db), current_user=Depends(oauth2.get_current_user)):
     post = db.get(models.Post, id)
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"Post with id: {id} is not found")
+    if (current_user.id != post.owner_id):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f"Not authorized to update this post")
     for field, value in post_data.model_dump().items():
         setattr(post, field, value)
     db.commit()
